fix: load events as soon as Gatekeeper is ready
